requirements.txt
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
debtcollector==1.22.0
docopt==0.6.2
flufl.enum==4.1.1
hupper==1.9.1
idna==2.8
importlib-metadata==1.4.0
iso8601==0.1.12
Jinja2==2.10.3
jsonschema==3.2.0
keystoneauth1==3.18.0
lxml==4.4.2
MarkupSafe==1.1.1
more-itertools==8.1.0
msgpack==0.6.2
netaddr==0.7.19
netifaces==0.10.9
numpy==1.18.1
os-service-types==1.7.0
oslo.config==7.0.0
oslo.i18n==3.25.1
oslo.serialization==2.29.2
oslo.utils==3.42.1
OWSLib==0.19.0
Paste==3.2.6
PasteDeploy==2.0.1
pathlib==1.0.1
pbr==5.4.4
plaster==1.0
plaster-pastedeploy==0.7
Pydap==3.2.2
pyparsing==2.4.6
pyproj==2.4.2.post1
pyramid==1.10.4
pyrsistent==0.15.7
python-dateutil==2.8.1
python-keystoneclient==3.22.0
python-swiftclient==3.8.1
pytz==2019.3
pywps==4.2.3
PyYAML==5.3
requests==2.22.0
rfc3986==1.3.2
# Editable Git install with no remote (silvereye-wps-demo==0.1)
-e /Users/diego/Documents/code/eRes/silvereye_wps_demo/src
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
stevedore==1.31.0
translationstring==1.3
urllib3==1.25.7
venusian==3.0.0
waitress==1.4.2
WebOb==1.8.5
Werkzeug==0.16.0
wrapt==1.11.2
zipp==1.0.0
zope.deprecation==4.4.0
zope.interface==4.7.1
